# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIBE Restaurant Ordering API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  category?: string;
  imageUrl?: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  done: 'done',
} as const;

export interface Order {
  id: number;
  orderNumber: number;
  status: OrderStatus;
  totalAmount: number;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  done: 'done',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface DashboardSummary {
  totalSalesToday: number;
  totalOrdersToday: number;
  totalOrdersAllTime: number;
  /** @nullable */
  bestSellingItem: string | null;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  done: 'done',
} as const;