# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIBE Restaurant Ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSalesToday: number;
  totalOrdersToday: number;
  totalOrdersAllTime: number;
  /** @nullable */
  bestSellingItem: string | null;
}