# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIBE Restaurant Ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}