# lib/api-zod/src/generated/types/topProduct.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VIBE Restaurant Ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}